MaryiaF/refactor: the hierarchy structure of journal
