--- packages/bot-web-ui/src/components/journal/filter-dialog/index.js
@@ -1,3 +1,0 @@
-import { FilterDialog } from './filter-dialog';
-
-export default FilterDialog;
