+++ packages/bot-web-ui/src/components/journal/index.ts
