+++ packages/bot-web-ui/src/components/journal/journal-components/date-item.tsx
@@ -1,8 +1,10 @@
 import React from 'react';
 import { TDateItemProps } from '../journal.types';
 
-export const DateItem = ({ date, time }: TDateItemProps) => (
+const DateItem = ({ date, time }: TDateItemProps) => (
     <>
         <span className='journal__text-date'>{date}</span> | <span className='journal__text-time'>{time}</span>
     </>
 );
+
+export default DateItem;
