--- packages/bot-web-ui/src/components/journal/journal-item/index.js
@@ -1,3 +1,0 @@
-import { JournalItem } from './journal-item';
-
-export default JournalItem;
