--- packages/bot-web-ui/src/components/journal/journal-loader/index.js
@@ -1,3 +1,0 @@
-import { JournalLoader } from './journal-loader';
-
-export default JournalLoader;
