--- packages/bot-web-ui/src/components/journal/journal-tools/index.js
@@ -1,3 +1,0 @@
-import { JournalTools } from './journal-tools';
-
-export default JournalTools;
